refactor: move composer Alpine state into data module
