+++ resources/js/app.js
@@ -72,4 +72,7 @@ Alpine.data('themeSwitch', themeSwitch);
 import { poll } from './poll.js';
 Alpine.data('poll', poll);
 
+import { questionComposer } from './question-composer.js';
+Alpine.data('questionComposer', questionComposer);
+
 Livewire.start()
