+++ resources/views/livewire/questions/create.blade.php
@@ -17,209 +17,7 @@
         wire:keydown.ctrl.enter="store"
         data-post-composer
         data-draft-key="{{ $this->draftKey }}"
-        x-on:post-modal-poll.window="
-            if ($wire.customDraftKey === 'post_modal') {
-                isPoll = $event.detail.isPoll;
-                pollOptions = $event.detail.pollOptions;
-                pollDuration = $event.detail.pollDuration;
-                threadPolls = $event.detail.threadPolls;
-                images = $event.detail.images;
-            }
-        "
-        x-data="{
-            ...imageUpload(),
-            ...poll(),
-            initComponents() {
-                imageUpload().init.call(this);
-                poll().init.call(this);
-            },
-            content: $persist($wire.entangle('content')).as('{{ $this->draftKey }}'),
-            threadPosts: $persist($wire.entangle('threadPosts')).as('{{ $this->draftKey }}_posts'),
-            threadPolls: $persist($wire.entangle('threadPolls')).as('{{ $this->draftKey }}_polls'),
-            ensureThreadPolls() {
-                while (this.threadPolls.length < this.threadPosts.length) {
-                    this.threadPolls.push(this.emptyThreadPoll());
-                }
-
-                this.threadPolls.splice(this.threadPosts.length);
-            },
-            hasDraft() {
-                if ((this.content || '').trim() !== '') {
-                    return true;
-                }
-
-                return (this.threadPosts || []).some((post) => (post || '').trim() !== '');
-            },
-            discardDraft() {
-                (this.images || []).forEach((image) => {
-                    this.$wire.deleteImageAfterValidation(this.normalizePath(image.path));
-                });
-
-                this.$wire.$errors.clear();
-                this.content = '';
-                this.threadPosts = [];
-                this.threadPolls = [];
-                this.images = [];
-                this.removeErrors();
-            },
-            canAddPost() {
-                if (this.uploading) {
-                    return false;
-                }
-
-                if ((this.content || '').trim() === '') {
-                    return false;
-                }
-
-                return (this.threadPosts || []).every((post) => (post || '').trim() !== '');
-            },
-            addPost() {
-                if (this.uploading) {
-                    return;
-                }
-
-                if (this.$wire.customDraftKey !== 'post_modal') {
-                    this.continueToModal();
-
-                    return;
-                }
-
-                if (! this.canAddPost()) {
-                    this.focusFirstEmptyPost();
-
-                    return;
-                }
-
-                if (this.threadPosts.length < {{ $this->maxThreadPosts - 1 }}) {
-                    this.threadPosts.push('');
-                    this.threadPolls.push(this.emptyThreadPoll());
-
-                    this.focusLastPost();
-                }
-            },
-            focusFirstEmptyPost() {
-                if ((this.content || '').trim() === '') {
-                    this.$refs.content?.focus();
-
-                    return;
-                }
-
-                this.$nextTick(() => {
-                    const textareas = this.$root.querySelectorAll('[data-thread-post]');
-                    const index = (this.threadPosts || []).findIndex((post) => (post || '').trim() === '');
-
-                    textareas[index]?.focus();
-                });
-            },
-            continueToModal() {
-                const content = (this.content || '').trim();
-                const threadPosts = (this.threadPosts || []).filter((post) => (post || '').trim() !== '');
-                const pollOptions = [...this.pollOptions];
-                const threadPolls = this.threadPolls.map((poll) => ({
-                    ...poll,
-                    options: [...poll.options],
-                }));
-                const images = this.images.map((image) => ({ ...image }));
-
-                if (content !== '' || threadPosts.length > 0) {
-                    this.$wire.$errors.clear();
-                    this.$wire.dispatch('thread.continue-in-modal', {
-                        content: content,
-                        threadPosts: threadPosts,
-                        threadPolls: threadPolls,
-                        isPoll: Boolean(this.isPoll),
-                        pollOptions: pollOptions,
-                        pollDuration: this.pollDuration,
-                        images: images,
-                        sourceDraftKey: '{{ $this->draftKey }}',
-                    });
-                    window.dispatchEvent(new CustomEvent('post-modal-poll', {
-                        detail: {
-                            isPoll: Boolean(this.isPoll),
-                            pollOptions: pollOptions,
-                            pollDuration: this.pollDuration,
-                            threadPolls: threadPolls,
-                            images: images,
-                        },
-                    }));
-
-                    this.content = '';
-                    this.threadPosts = [];
-                    this.isPoll = false;
-                    this.pollOptions = ['', ''];
-                    this.pollDuration = 1;
-                }
-
-                this.$dispatch('open-modal', 'post-create');
-            },
-            removePost(index) {
-                this.images
-                    .filter((image) => image.target === index)
-                    .forEach((image) => {
-                        this.$wire.deleteImageAfterValidation(this.normalizePath(image.path));
-                    });
-
-                this.images = this.images
-                    .filter((image) => image.target !== index)
-                    .map((image) => image.target !== null && image.target > index
-                        ? { ...image, target: image.target - 1 }
-                        : image
-                    );
-
-                if (this.activeImageTarget === index) {
-                    this.activeImageTarget = null;
-                } else if (this.activeImageTarget !== null && this.activeImageTarget > index) {
-                    this.activeImageTarget -= 1;
-                }
-
-                this.threadPosts.splice(index, 1);
-                this.threadPolls.splice(index, 1);
-            },
-            emptyThreadPoll() {
-                return { isPoll: false, options: ['', ''], duration: 1 };
-            },
-            toggleThreadPoll(index) {
-                if (! this.threadPolls[index]) {
-                    this.threadPolls[index] = this.emptyThreadPoll();
-                }
-
-                this.threadPolls[index].isPoll = ! this.threadPolls[index].isPoll;
-            },
-            addThreadPollOption(index) {
-                if (! this.threadPolls[index]) {
-                    this.threadPolls[index] = this.emptyThreadPoll();
-                }
-
-                if (this.threadPolls[index].options.length < 4) {
-                    this.threadPolls[index].options.push('');
-                }
-            },
-            removeThreadPollOption(index, optionIndex) {
-                if (this.threadPolls[index].options.length > 2) {
-                    this.threadPolls[index].options.splice(optionIndex, 1);
-                }
-            },
-            focusLastPost() {
-                this.$nextTick(() => {
-                    const textareas = this.$root.querySelectorAll('[data-thread-post]');
-                    const last = textareas[textareas.length - 1];
-
-                    if (! last) {
-                        return;
-                    }
-
-                    last.focus();
-                    last.scrollIntoView({ behavior: 'smooth', block: 'nearest' });
-                });
-            },
-        }"
-        x-init='() => {
-            uploadLimit = {{ $this->uploadLimit }};
-            maxFileSize = {{ $this->maxFileSize }};
-            maxContentLength = {{ $this->maxContentLength }};
-            initComponents();
-            ensureThreadPolls();
-        }'
+        x-data="questionComposer({ draftKey: '{{ $this->draftKey }}', maxThreadPosts: {{ $this->maxThreadPosts }}, uploadLimit: {{ $this->uploadLimit }}, maxFileSize: {{ $this->maxFileSize }}, maxContentLength: {{ $this->maxContentLength }} })"
         class="{{ $isModalComposer ? 'flex min-h-0 flex-1 flex-col' : '' }} pb-0"
     >
         <div class="{{ $isModalComposer ? 'min-h-0 flex-1 overflow-y-auto' : '' }}">
